refactor(devtools): share relay installer
