+++ packages/inspector/src/chrome-extension/content-script.ts
@@ -1,0 +1,6 @@
+import {
+  createDevtoolsPortRelayConfig,
+  installDevtoolsPortRelay,
+} from "jazz-tools/dev-tools/relay-installer";
+
+installDevtoolsPortRelay(createDevtoolsPortRelayConfig());
